add magicgui.pyi
