+++ setup.cfg
@@ -112,12 +112,6 @@ show_column_numbers = True
 show_error_codes = True
 # pretty = True
 
-[mypy-magicgui._tests.*]
-ignore_errors = True
-
-[mypy-magicgui._qt._tests.*]
-ignore_errors = True
-
 [mypy-.examples/]
 ignore_errors = True
